🧹 remove slack query that is not useful (#332)
After quite some consideration we decided to remove the check for
whitelisted domains since it is not adding a lot of security value. We
already check for external members and match it with the channel name.
The query is also very slow for large slack accounts (caused by slack
api rate limiting).